Separate the hook in a dedicated file
