+++ src/components/react-hook-form/ExpandingTextField.tsx
@@ -9,7 +9,7 @@ import { FunctionComponent, useState } from 'react';
 import { TextFieldProps, Theme, Typography } from '@mui/material';
 import { useWatch } from 'react-hook-form';
 import { TextInput, TextInputProps } from '../..';
-import { useCustomFormContext } from './custom-form-provider.tsx';
+import { useCustomFormContext } from './provider/use-custom-form-context';
 
 interface ExpandingTextFieldProps extends TextInputProps {
     name: string;
