+++ src/components/react-hook-form/autocomplete-input.jsx
@@ -15,7 +15,7 @@ import {
     isFieldRequired,
 } from './utils/functions';
 import FieldLabel from './utils/field-label';
-import { useCustomFormContext } from './custom-form-provider';
+import { useCustomFormContext } from './provider/use-custom-form-context';
 
 const AutocompleteInput = ({
     name,
