+++ src/index.d.ts
@@ -104,10 +104,8 @@ export { default as directory_items_input_fr } from './components/translations/d
 export { TagRenderer } from './components/ElementSearchDialog';
 export { EquipmentItem } from './components/ElementSearchDialog/equipment-item';
 export { useIntlRef } from './hooks/useIntlRef';
-export {
-    default as CustomFormProvider,
-    useCustomFormContext,
-} from './components/react-hook-form/custom-form-provider';
+export { useCustomFormContext } from './components/react-hook-form/provider/use-custom-form-context';
+export { default as CustomFormProvider } from './components/react-hook-form/provider/custom-form-provider';
 export { default as SliderInput } from './components/react-hook-form/slider-input';
 export { default as TextFieldWithAdornment } from './components/react-hook-form/utils/text-field-with-adornment';
 export {
