+++ src/index.js
@@ -96,10 +96,8 @@ export { default as CardErrorBoundary } from './components/CardErrorBoundary';
 export { useIntlRef } from './hooks/useIntlRef';
 export { useSnackMessage } from './hooks/useSnackMessage';
 export { useDebounce } from './hooks/useDebounce';
-export {
-    default as CustomFormProvider,
-    useCustomFormContext,
-} from './components/react-hook-form/custom-form-provider';
+export { useCustomFormContext } from './components/react-hook-form/provider/use-custom-form-context';
+export { default as CustomFormProvider } from './components/react-hook-form/provider/custom-form-provider';
 export { default as AutocompleteInput } from './components/react-hook-form/autocomplete-input';
 export { default as TextInput } from './components/react-hook-form/text-input';
 export { default as ExpandingTextField } from './components/react-hook-form/ExpandingTextField';
